Use libsrtp 1.4.4
Consistent with existing builds and production environment. The required version has been added to the mcu-release yum repo, as it is no longer available from epel.

Once this version is deployed it should be trivial to update libsrtp to a newer version in a future release, as yum should take care of upgrades/downgrades of the libsrtp dependency when upgrading/downgrading the MCU.

JIRA: WEBRTC-1874