Allow for merging, rather than overwriting, of candidates when SDP comes in.
This should maintain statistics on STUN requests when candidate is created "on the fly" before an answer has been returned on our offer. It also brings the code slightly closer to trickle ICE support. Finally, it removes and invalid "short circuit" that was being done when the answer came in, going directly to the e_Completed state. This was not often being triggered, and other changes means it could recover, but it is still incorrect and was removed.