Fixed possible memory leak in RTCP extended metrics.
If the JItter Buffer is disabled, then the extended metrics calculation grows a list for every packet. This cause ememory and CPU usage to go up over time.

Fix is done in two places, first prevent the trigger of zero JB delay. Second, put a maximum on the size of the lists as a fail safe.

Also did some code cleaning up.